requirements.txt
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.1.8
Flask==3.1.3
gunicorn==23.0.0
idna==3.11
importlib_metadata==8.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
numpy==2.0.2
packaging==26.0
pandas==2.3.3
pycryptodome==3.23.0
python-dateutil==2.9.0.post0
pytz==2026.1.post1
requests==2.32.5
six==1.17.0
tzdata==2026.1
urllib3==2.6.3
Werkzeug==3.1.8
zipp==3.23.0
joblib==1.4.2